Move JWT auth middleware to faas-middleware package
