+++ go.mod
@@ -4,11 +4,9 @@ go 1.21
 
 require (
 	github.com/docker/go-units v0.5.0
-	github.com/golang-jwt/jwt/v5 v5.2.1
-	github.com/openfaas/faas-middleware v1.2.3
+	github.com/openfaas/faas-middleware v1.2.4
 	github.com/openfaas/faas-provider v0.25.3
 	github.com/prometheus/client_golang v1.19.0
-	github.com/rakutentech/jwk-go v1.1.3
 )
 
 require (
@@ -22,6 +20,8 @@ require (
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/cespare/xxhash/v2 v2.3.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/golang-jwt/jwt/v5 v5.2.1 // indirect
+	github.com/rakutentech/jwk-go v1.1.3 // indirect
 	golang.org/x/crypto v0.22.0 // indirect
 	google.golang.org/protobuf v1.33.0 // indirect
 )
